refactor: Hoist `hook` to `src` directory
Align with Gutenberg project practices.
